feat: :sparkles: Introduce new type definitions for Order, Project, ProjectTree, Species, and User; refactor existing types for improved organization and clarity
