+++ src/@types/Project.d.ts
@@ -1,0 +1,12 @@
+export interface IProject {
+  id: number;
+  name: string;
+  description: string;
+  status: string;
+  picture: string;
+  city: string;
+  country: string;
+  continent: string;
+  createdAt?: string;
+  updatedAt?: string;
+}
